# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
  /model3d/capabilities:
    get:
      tags: [models]
      summary: 读取生效的 3D 模型生成能力
      description: 返回 available、providerName、policyRevision、activeConfigVersion、defaultModel、modelVersions、modes 和 inputLimits，不返回凭据。无生效配置时 available=false。
      responses:
        "200": {$ref: "#/components/responses/Envelope"}
        "401": {$ref: "#/components/responses/Error"}
  /model3d/tasks:
    post:
      tags: [tasks]
      summary: 创建文本、单图或多视图 3D 任务
      description: requestId 在当前用户内幂等；相同编号但内容不同返回冲突。冻结生效配置与输入资源版本。上游创建只发送一次，不自动重发结果不确定的请求。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Model3DCreateRequest"}
      responses:
        "200": {description: 标准信封 data 为 Model3DTaskView}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /model3d/tasks/{id}:
    get:
      tags: [tasks]
      summary: 读取本人 3D 任务和平台模型结果
      parameters:
        - {in: path, name: id, required: true, schema: {type: string}}
      responses:
        "200": {description: 标准信封 data 为 Model3DTaskView}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /model3d/requests/{requestId}:
    get:
      tags: [tasks]
      summary: 创建响应丢失后按本人请求编号只读恢复
      description: 未找到返回 data=null，此接口不会创建任务或调用生成接口。
      parameters:
        - {in: path, name: requestId, required: true, schema: {type: string}}
      responses:
        "200": {description: 标准信封 data 为 Model3DTaskView 或 null}
        "401": {$ref: "#/components/responses/Error"}
  /model3d/tasks/{id}/recover:
    post:
      tags: [tasks]
      summary: 恢复已有 3D 上游任务的模型回存
      description: 仅允许恢复已有上游任务 ID 的可恢复失败，不创建新的供应商任务。
      parameters:
        - {in: path, name: id, required: true, schema: {type: string}}
      responses:
        "200": {description: 标准信封 data 为 Model3DTaskView}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /admin/model3d:
    get:
      tags: [admin]
      summary: 获取 3D 服务商配置与生效策略
      description: 仅管理员。返回 providers、providerTypes、activeProviderId、activeConfigId、policyRevision；凭据仅返回 apiKeyConfigured。
      responses:
        "200": {$ref: "#/components/responses/Envelope"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
  /admin/model3d/providers:
    post:
      tags: [admin]
      summary: 创建 3D 服务商与首个配置版本
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Model3DProviderRequest"}
      responses:
        "200": {description: 标准信封 data 为服务商配置视图，不含 API Key}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
  /admin/model3d/providers/{id}:
    parameters:
      - {in: path, name: id, required: true, schema: {type: string}}
    put:
      tags: [admin]
      summary: 保存 3D 服务商的新配置版本
      description: 不自动切换生效版本。API Key 留空保留原凭据，类型不可更改。保存不请求云端。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Model3DProviderRequest"}
      responses:
        "200": {description: 标准信封 data 为服务商配置视图}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
    delete:
      tags: [admin]
      summary: 归档非生效的 3D 服务商配置
      description: 保留历史模型、任务与配置版本。
      responses:
        "200": {$ref: "#/components/responses/Envelope"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /admin/model3d/providers/{id}/activate:
    post:
      tags: [admin]
      summary: 按策略修订激活指定 3D 配置版本
      parameters:
        - {in: path, name: id, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [configId, expectedRevision]
              properties:
                configId: {type: string}
                expectedRevision: {type: integer, format: int64, minimum: 0}
      responses:
        "200": {$ref: "#/components/responses/Envelope"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /admin/model3d/disable:
    post:
      tags: [admin]
      summary: 停用新的 3D 生成任务
      description: 在途任务继续使用受理时的旧配置，不重复提交上游。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedRevision]
              properties:
                expectedRevision: {type: integer, format: int64, minimum: 0}
      responses:
        "200": {$ref: "#/components/responses/Envelope"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
    Model3DProviderRequest:
      type: object
      required: [name, type, enabled, defaultModel, allowedModels, allowedModes, timeoutSeconds, maxTasksPerDay]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        type: {type: string, enum: [tripo3d]}
        enabled: {type: boolean}
        defaultModel: {type: string}
        allowedModels: {type: array, minItems: 1, items: {type: string}}
        allowedModes: {type: array, minItems: 1, items: {type: string, enum: [text, image, multiview]}}
        timeoutSeconds: {type: integer, minimum: 10, maximum: 120}
        maxTasksPerDay: {type: integer, minimum: 1, maximum: 100000}
        apiKey: {type: string, writeOnly: true, description: 首次必填，后续留空保留已有密钥}
    Model3DParameters:
      type: object
      additionalProperties: false
      required: [model, texture, pbr]
      properties:
        model: {type: string, enum: [v3.1-20260211, v3.0-20250812, v2.5-20250123]}
        negativePrompt: {type: string, maxLength: 255}
        texture: {type: boolean, description: 必须显式传入布尔值，缺失或 null 拒绝}
        pbr: {type: boolean, description: 必须显式传入布尔值，开启时必须同时开启 texture}
        faceLimit: {type: integer, minimum: 1, description: 不传为自适应，上限按模型版本、质量、四边面和低模参数组合校验}
        modelSeed: {type: integer}
        imageSeed: {type: integer}
        textureSeed: {type: integer}
        textureQuality: {type: string, enum: [standard, detailed, extreme, fast], description: 仅 H3.0/H3.1，fast 必须使用贴图版本 v3.5-20260815}
        textureVersion: {type: string, enum: [v2.5-20250123, v3.0-20250812, v3.5-20260815]}
        geometryQuality: {type: string, enum: [standard, detailed]}
        autoSize: {type: boolean}
        quad: {type: boolean, description: 四边面会产生 FBX 格式}
        smartLowPoly: {type: boolean}
        compress: {type: string, enum: ["", geometry]}
        exportOrientation: {type: string, enum: ["+x", "-x", "+y", "-y"]}
        exportUv: {type: boolean, default: true}
        delight: {type: boolean, default: true, description: 仅 texture 开启且 textureVersion 为 v3.5-20260815 时可传}
        enableImageAutofix: {type: boolean, default: false, description: 仅单图模式可传}
        textureAlignment: {type: string, enum: [original_image, geometry]}
        orientation: {type: string, enum: [default, align_image]}
    Model3DCreateRequest:
      type: object
      additionalProperties: false
      required: [requestId, sourceFingerprint, canvasId, nodeId, mode, parameters, expectedPolicyRevision]
      properties:
        requestId: {type: string, description: 当前用户内唯一的请求标识，重试保持不变}
        sourceFingerprint: {type: string, description: 前端输入与参数快照身份，仅用于上下文回写，不替代后端请求摘要}
        canvasId: {type: string}
        nodeId: {type: string}
        mode: {type: string, enum: [text, image, multiview]}
        prompt: {type: string, maxLength: 1024}
        imageResourceId: {type: string}
        views:
          type: object
          additionalProperties: false
          required: [front]
          minProperties: 2
          maxProperties: 4
          properties:
            front: {type: string}
            left: {type: string}
            back: {type: string}
            right: {type: string}
        parameters: {$ref: "#/components/schemas/Model3DParameters"}
        expectedPolicyRevision: {type: integer, format: int64, minimum: 0}
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - upstream_dns_failed
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
...
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"